Add param initialization from resource finder
Add method in ParamHelperServer for initializing parameters from
ResourceFinder.
Add new type of IO for parameters, that is PARAM_CONFIG, for parameters
that are set at configuration time, but after that cannot be modified.
Reinsert memory allocation for parameters that are not linked to an
external variable (to not force user to initialize after all parameters
have been linked).