Invert order of torso joints
When reading encoders/pwm or sending commands to the motors (pos, vel,
torque, pwm) invert order of torso joints (i.e. joint 0 becomes joint 2
and the other way around, while joint 1 remains joint 1). This is due to
a peculiarity of iCub's structure and it is necessary to make the
sensors/motors structure match the kinematic/dynamic model of the robot.