[TorqueBalancing] fixes reference generator
This commit fixes a BIG issue with the smoothing reference generator: the Ts parameter was passed in milliseconds instead of seconds.

In the mean time it added support to obtain velocity and acceleration from the smoothing generator (currently not used).
Also added a constant reader.
Fixed some warnings
Fixed also an issue with the impedance generator: because the generator simply outputs a reference (so it is not a controller), when it is activated the smoother starts from 0.
Without smoothing this would not have been a problem, but the smoother in this way starts from zero and in T time it reaches 90% of the reference, which is not the desired behaviour.
I now added an option to force the generator to reset the filter to have the same value of the reference, instead of zero.

@DanielePucci