[wholeBodyDynamicsTree] Fix non initialized flags if SIMPLE_LEGGED_ODOMETRY was not present
Also, update configuration file of iCubGenova02 and icubGazeboSim to not include
the SIMPLE_LEGGED_ODOMETRY group. The SIMPLE_LEGGED_ODOMETRY was an hack for the codyco second year
demo, and the proper functionality will be implemented in the new wholeBodyEstimator module.
The wbi needs to be configured with the externalFloatingBaseStatePort option to read the port,
and no robot has that flag in their wbi configuration.